Backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-cors-headers==4.7.0
django-rest-knox==5.0.2
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
pillow==11.2.1
PyJWT==2.9.0
requests==2.32.3
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
